browser_agent/requirements.txt
# proxy pypi
--index-url https://mirrors.cloud.tencent.com/pypi/simple/
# private pypi
--extra-index-url https://mirrors.tencent.com/repository/pypi/tencent_pypi/simple/


# Install package need

wheel

# add the requirements for trpc project

# protobuf==3.20.3
PyYAML
aiohttp
netifaces==0.10.9
concurrent-log-asyncio==0.10.1
automaxprocs==2.0.0
Cython==0.29.21
six==1.15.0
pytest>=5.2.2
python-snappy>=0.5.4
uvloop>=0.10.2
python-gflags>=3.1.2
setproctitle
uv

# add for your project
# Add the stub package like this:
#
# git+http://git.woa.com/trpc-python/stubs/stub-package.git#egg=stub-package
# 
#
sseclient-py

# fastapi
fastapi
authlib
passlib
pyjwt
uvicorn
# fastembed
itsdangerous
python-multipart
mcp==1.13.1
fastapi-mcp==0.4.0


# gradio
gradio>5.0.0
modelscope-studio
openai

# openmanus
numpy
datasets>=3.4.1
tiktoken>=0.9.0

html2text>=2024.2.26
gymnasium>=1.1.1
pillow>=11.1.0
browsergym>=0.13.3
unidiff>=0.7.5
googlesearch-python>=1.3.0
baidusearch>=1.0.3
duckduckgo_search>=7.5.3

aiofiles>=24.1.0
pydantic_core>=2.27.2
colorama>=0.4.6

patchright==1.56.0
playwright==1.56.0

docker>=7.1.0
pytest-asyncio>=0.25.3

tomli>=2.0.0
boto3>=1.37.18
beautifulsoup4>=4.13.3

# browser_use
anyio>=4.9.0
httpx>=0.27.2
python-dotenv>=1.0.1
requests>=2.32.3
posthog>=3.7.0
patchright>=1.51.0
markdownify>=1.1.0
langchain-core>=0.3.49
langchain-openai>=0.3.11
langchain-anthropic>=0.3.3
langchain-ollama>=0.3.0
langchain-google-genai>=2.1.2
langchain-deepseek>=0.1.3
langchain>=0.3.21
langchain-aws>=0.2.11
google-api-core>=2.24.0
pyperclip>=1.9.0
pyobjc>=11.0; # platform_system == 'darwin'
# screeninfo>=0.8.1 # ; platform_system != 'darwin'
typing-extensions>=4.12.2
psutil>=7.0.0
faiss-cpu==1.9.0
mem0ai==0.1.93
rich>=14.0.0
click>=8.1.8
textual>=3.2.0
dotenv
